+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added the field layout designer to campaign types, mailing list types and contact layouts ([#163](https://github.com/putyourlightson/craft-campaign/issues/163), [#198](https://github.com/putyourlightson/craft-campaign/issues/198), [#269](https://github.com/putyourlightson/craft-campaign/issues/269)).
 - Added drafts and revisions to campaigns ([#301](https://github.com/putyourlightson/craft-campaign/issues/301)).
 - Added a "duplicate" action to campaigns ([#292](https://github.com/putyourlightson/craft-campaign/issues/292)).
+- Added condition settings to the Campaigns, Contacts and Mailing Lists relation fields.
 - Added campaign and mailing list type user group permissions.
 - Added the ability to view disabled campaigns using a token URL.
 - Added the `enableAnonymousTracking` setting to campaign types, which prevents tracking of personal data of contacts ([#115](https://github.com/putyourlightson/craft-campaign/issues/115)).